requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
elevenlabslib==0.4.1
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
pipenv==2023.2.18
platformdirs==3.0.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfiglet==0.8.post1
Pygments==2.15.1
pyparsing==3.0.9
python-dotenv==1.0.0
python-resize-image==1.1.20
requests==2.28.2
requests-cache==1.0.1
rich==13.3.4
rsa==4.9
six==1.16.0
tqdm==4.65.0
typing==3.7.4.3
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.14
virtualenv==20.20.0
virtualenv-clone==0.5.7
yarl==1.8.2
